+++ samples/firefox-accounts/src/main/java/org/mozilla/samples/fxa/MainActivity.kt
@@ -4,7 +4,6 @@
 
 package org.mozilla.samples.fxa
 
-import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
@@ -14,19 +13,24 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.browser.customtabs.CustomTabsIntent
 import androidx.core.app.ActivityCompat
-import androidx.core.content.edit
 import androidx.core.net.toUri
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
-import mozilla.appservices.fxaclient.FxaConfig
 import mozilla.appservices.fxaclient.FxaServer
+import mozilla.components.concept.sync.AccountObserver
+import mozilla.components.concept.sync.AuthType
+import mozilla.components.concept.sync.DeviceCapability
+import mozilla.components.concept.sync.DeviceConfig
+import mozilla.components.concept.sync.DeviceType
+import mozilla.components.concept.sync.OAuthAccount
 import mozilla.components.concept.sync.Profile
 import mozilla.components.feature.qr.QrFeature
 import mozilla.components.lib.fetch.httpurlconnection.HttpURLConnectionClient
-import mozilla.components.service.fxa.FirefoxAccount
-import mozilla.components.service.fxa.FxaException
+import mozilla.components.service.fxa.FxaAuthData
+import mozilla.components.service.fxa.ServerConfig
+import mozilla.components.service.fxa.manager.FxaAccountManager
 import mozilla.components.support.base.log.Log
 import mozilla.components.support.base.log.sink.AndroidLogSink
 import mozilla.components.support.ktx.android.view.setupPersistentInsets
@@ -35,7 +39,6 @@ import mozilla.components.support.rustlog.RustLog
 import kotlin.coroutines.CoroutineContext
 
 open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteListener, CoroutineScope {
-    private lateinit var account: FirefoxAccount
     private var scopesWithoutKeys: Set<String> = setOf("profile")
     private var scopesWithKeys: Set<String> = setOf("profile", "https://identity.mozilla.com/apps/oldsync")
     private var scopes: Set<String> = scopesWithoutKeys
@@ -46,12 +49,24 @@ open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteList
     override val coroutineContext: CoroutineContext
         get() = Dispatchers.Main + job
 
+    private val accountManager by lazy {
+        FxaAccountManager(
+            context = applicationContext,
+            serverConfig = ServerConfig(FxaServer.Custom(CONFIG_URL), CLIENT_ID, REDIRECT_URL),
+            deviceConfig = DeviceConfig(
+                name = "A-C FxA Sample",
+                type = DeviceType.MOBILE,
+                capabilities = setOf(DeviceCapability.SEND_TAB),
+            ),
+            syncConfig = null,
+            applicationScopes = scopesWithKeys,
+        )
+    }
+
     companion object {
         const val CLIENT_ID = "3c49430b43dfba77"
         const val CONFIG_URL = "https://accounts.firefox.com"
         const val REDIRECT_URL = "$CONFIG_URL/oauth/success/3c49430b43dfba77"
-        const val FXA_STATE_PREFS_KEY = "fxaAppState"
-        const val FXA_STATE_KEY = "fxaState"
         private const val REQUEST_CODE_CAMERA_PERMISSIONS = 1
     }
 
@@ -67,7 +82,9 @@ open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteList
         Log.addSink(AndroidLogSink())
 
         job = Job()
-        account = initAccount()
+
+        accountManager.register(accountObserver, owner = this, autoPause = true)
+        launch { accountManager.start() }
 
         qrFeature = QrFeature(
             this,
@@ -77,7 +94,11 @@ open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteList
             },
             onScanResult = { pairingUrl ->
                 launch {
-                    val url = account.beginPairingFlow(pairingUrl, scopes, SampleFxAEntryPoint.HomeMenu)
+                    val url = accountManager.beginAuthentication(
+                        pairingUrl = pairingUrl,
+                        entrypoint = SampleFxAEntryPoint.HomeMenu,
+                        authScopes = scopes,
+                    )
                     if (url == null) {
                         Log.log(
                             Log.Priority.ERROR,
@@ -86,7 +107,7 @@ open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteList
                         )
                         return@launch
                     }
-                    openWebView(url.url)
+                    openWebView(url)
                 }
             },
             scanMessage = R.string.pair_instructions_message,
@@ -96,17 +117,13 @@ open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteList
 
         findViewById<View>(R.id.buttonCustomTabs).setOnClickListener {
             launch {
-                account.beginOAuthFlow(scopes, SampleFxAEntryPoint.HomeMenu)?.let {
-                    openTab(it.url)
-                }
+                beginAuthentication()?.let { openTab(it) }
             }
         }
 
         findViewById<View>(R.id.buttonWebView).setOnClickListener {
             launch {
-                account.beginOAuthFlow(scopes, SampleFxAEntryPoint.HomeMenu)?.let {
-                    openWebView(it.url)
-                }
+                beginAuthentication()?.let { openWebView(it) }
             }
         }
 
@@ -115,11 +132,7 @@ open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteList
         }
 
         findViewById<View>(R.id.buttonLogout).setOnClickListener {
-            getSharedPreferences(FXA_STATE_PREFS_KEY, Context.MODE_PRIVATE).edit {
-                putString(FXA_STATE_KEY, "")
-            }
-            val txtView: TextView = findViewById(R.id.txtView)
-            txtView.text = getString(R.string.logged_out)
+            launch { accountManager.logout() }
         }
 
         findViewById<CheckBox>(R.id.checkboxKeys).setOnCheckedChangeListener { _, isChecked ->
@@ -127,23 +140,12 @@ open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteList
         }
     }
 
-    private fun initAccount(): FirefoxAccount {
-        getAuthenticatedAccount()?.let {
-            launch {
-                it.getProfile(true)?.let { profile ->
-                    displayProfile(profile)
-                }
-            }
-            return it
-        }
-
-        val config = FxaConfig(FxaServer.Custom(CONFIG_URL), CLIENT_ID, REDIRECT_URL)
-        return FirefoxAccount(config)
-    }
+    private suspend fun beginAuthentication(): String? =
+        accountManager.beginAuthentication(entrypoint = SampleFxAEntryPoint.HomeMenu, authScopes = scopes)
 
     override fun onDestroy() {
         super.onDestroy()
-        account.close()
+        accountManager.close()
         job.cancel()
     }
 
@@ -156,23 +158,18 @@ open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteList
             val url = data.toUri()
             val code = url.getQueryParameter("code")!!
             val state = url.getQueryParameter("state")!!
-            displayAndPersistProfile(code, state)
+            finishAuthentication(code, state)
         }
     }
 
     override fun onLoginComplete(code: String, state: String, fragment: LoginFragment) {
-        displayAndPersistProfile(code, state)
+        finishAuthentication(code, state)
         supportFragmentManager.popBackStack()
     }
 
-    private fun getAuthenticatedAccount(): FirefoxAccount? {
-        val savedJSON = getSharedPreferences(FXA_STATE_PREFS_KEY, Context.MODE_PRIVATE).getString(FXA_STATE_KEY, "")
-        return savedJSON?.let {
-            try {
-                FirefoxAccount.fromJSONString(it, null)
-            } catch (e: FxaException) {
-                null
-            }
+    private fun finishAuthentication(code: String, state: String) {
+        launch {
+            accountManager.finishAuthentication(FxaAuthData(AuthType.Signin, code = code, state = state))
         }
     }
 
@@ -194,23 +191,26 @@ open class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteList
         }
     }
 
-    private fun displayAndPersistProfile(code: String, state: String) {
-        launch {
-            account.completeOAuthFlow(code, state)
-            account.getProfile()?.let {
-                displayProfile(it)
-            }
-            account.toJSONString().let {
-                getSharedPreferences(FXA_STATE_PREFS_KEY, Context.MODE_PRIVATE).edit {
-                    putString(FXA_STATE_KEY, it)
-                }
-            }
-        }
-    }
-
     private fun displayProfile(profile: Profile) {
         val txtView: TextView = findViewById(R.id.txtView)
         txtView.text = getString(R.string.signed_in, "${profile.displayName ?: ""} ${profile.email}")
+    }
+
+    private val accountObserver = object : AccountObserver {
+        override fun onAuthenticated(account: OAuthAccount, authType: AuthType) {
+            launch {
+                account.getProfile()?.let { displayProfile(it) }
+            }
+        }
+
+        override fun onProfileUpdated(profile: Profile) {
+            displayProfile(profile)
+        }
+
+        override fun onLoggedOut() {
+            val txtView: TextView = findViewById(R.id.txtView)
+            txtView.text = getString(R.string.logged_out)
+        }
     }
 
     override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<String>, grantResults: IntArray) {